Esplora mode - Update nginx-mempool.conf
"""
Fought with this same issue for a while.

It turns out when you use the esplora mode, mempool.space itself will not serve the apis that are also served by electrs. Those apis are everything at the /api endpoint, whereas electrs uses /api/v1. The example nginx config essentially passes through /api to /api/v1, but that's incorrect for electrs.
"""
https://bitcointalk.org/index.php?topic=5457969.0